Update fwc_structure.py
Added a new parameter called self.global_offset to shift the complete plane along the x axis. The value is calculated from the wing root chord in order to make sure that the leading edge at the root does not pierce into the velocity field in front of the plane to alleviate sharp gradients due to periodicity.